--- engula/src/job/job.rs
@@ -1,9 +1,0 @@
-use super::compaction::{CompactionInput, CompactionOutput};
-
-pub enum JobInput {
-    Compaction(CompactionInput),
-}
-
-pub enum JobOutput {
-    Compaction(CompactionOutput),
-}
