+++ engula/src/job/local_runtime.rs
@@ -4,13 +4,13 @@ use async_trait::async_trait;
 use tokio::task;
 
 use super::compaction::{CompactionInput, CompactionOutput, FileMeta};
-use super::job::{JobInput, JobOutput};
 use super::runtime::JobRuntime;
 use crate::error::Result;
 use crate::file_system::FileSystem;
 use crate::format::{
     Iterator, MergingIterator, SstBuilder, SstOptions, SstReader, TableBuilder, TableReader,
 };
+use crate::job::{JobInput, JobOutput};
 
 pub struct LocalJobRuntime {
     fs: Arc<Box<dyn FileSystem>>,
