+++ engula/src/job/mod.rs
@@ -1,9 +1,15 @@
 mod compaction;
-mod job;
 mod local_runtime;
 mod runtime;
 
 pub use compaction::{CompactionInput, CompactionOutput, FileMeta};
-pub use job::{JobInput, JobOutput};
 pub use local_runtime::LocalJobRuntime;
 pub use runtime::JobRuntime;
+
+pub enum JobInput {
+    Compaction(CompactionInput),
+}
+
+pub enum JobOutput {
+    Compaction(CompactionOutput),
+}
