--- engula/src/journal/journal.rs
@@ -1,8 +1,0 @@
-use async_trait::async_trait;
-
-use crate::error::Result;
-
-#[async_trait]
-pub trait Journal: Send + Sync {
-    async fn append(&self, data: Vec<u8>) -> Result<()>;
-}
