+++ engula/src/memtable/btree_table.rs
@@ -8,8 +8,8 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use tokio::sync::{OwnedRwLockReadGuard, RwLock};
 
-use super::memtable::{MemItem, MemIter, MemSnapshot, MemTable};
 use crate::common::Timestamp;
+use crate::memtable::{MemItem, MemIter, MemSnapshot, MemTable};
 
 type Sequence = Reverse<Timestamp>;
 type SequenceTree = BTreeMap<Sequence, Vec<u8>>;
